# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Dieta e Treino API",
        "description": "API de dieta e treino desenvolvida em Laravel 12",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "api/auth": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticação do usuário, retornando seu token JWT",
                "operationId": "f58a21374d773f34d10a39a340945b5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "teste@phpunit.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "example": "senha"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido, retorna token JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "86400"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O e-mail não foi informado ou está inválido"
                                                    }
                                                },
                                                "senha": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Senha não foi informada ou incorreta (mínimo de 5 caracteres)"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Root"
                ],
                "summary": "Endpoint raiz da API",
                "description": "Retorna uma mensagem de teste 'Hello World'. Útil para verificar se a API está no ar.",
                "operationId": "d2f724c57f40ef229bad56a9f1d6ba6d",
                "responses": {
                    "200": {
                        "description": "Requisição executada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Hello World"
                            }
                        }
                    }
                }
            }
        },
        "api/user": {
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Cadastro do Usuário",
                "description": "Cadastro completo de um usuário no sistema.",
                "operationId": "04156b356a4bc86d53ff9b65d8d8d2bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Teste OpenAPI"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste_openapi@email.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "example": "senha_openapi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Teste OpenAPI"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "teste_openapi@email.com"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Pendente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Credenciais inválidas: o e-mail informado já possui cadastro no sistema."
                                                    }
                                                },
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo 'nome' é obrigatório."
                                                    }
                                                },
                                                "senha": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo 'senha' é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "put": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualizar usuário logado",
                "operationId": "e18206ec4fc17a4c90e24b4e64cbe6ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário logado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Teste OpenAPI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Teste OpenAPI"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "teste_openapi@email.com"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Pendente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Campo 'Nome' possui valor invalido com mais de 60 caracteres."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/confirm/{hash}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Confirmar cadastro do usuário",
                "operationId": "c44d90943497647c65ecd9a2174682e2",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash enviado por e-mail para confirmação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmado o cadastro do seu usuário. Agora você pode acessar a API!"
                    },
                    "404": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "Root",
            "description": "Root"
        },
        {
            "name": "Usuário",
            "description": "Usuário"
        }
    ]
}